Improve Callbacks
- Move callbacks to separate module
- Add EarlyStopping
- Move JSON logger
- Update history callback and jsoncallback
